fix warnings - precision loss

+++ xbmc/addons/kodi-dev-kit/include/kodi/tools/StringUtils.h
@@ -2945,19 +2945,19 @@ public:
       strHMS += StringUtils::Format("%i", seconds % 3600 / 60);
     else
     {
-      int hh = seconds / 3600;
+      long hh = seconds / 3600;
       seconds = seconds % 3600;
-      int mm = seconds / 60;
+      long mm = seconds / 60;
       int ss = seconds % 60;
 
       if (format == TIME_FORMAT_GUESS)
         format = (hh >= 1) ? TIME_FORMAT_HH_MM_SS : TIME_FORMAT_MM_SS;
       if (format & TIME_FORMAT_HH)
-        strHMS += StringUtils::Format("%2.2i", hh);
+        strHMS += StringUtils::Format("%2.2ld", hh);
       else if (format & TIME_FORMAT_H)
-        strHMS += StringUtils::Format("%i", hh);
+        strHMS += StringUtils::Format("%ld", hh);
       if (format & TIME_FORMAT_MM)
-        strHMS += StringUtils::Format(strHMS.empty() ? "%2.2i" : ":%2.2i", mm);
+        strHMS += StringUtils::Format(strHMS.empty() ? "%2.2ld" : ":%2.2ld", mm);
       if (format & TIME_FORMAT_SS)
         strHMS += StringUtils::Format(strHMS.empty() ? "%2.2i" : ":%2.2i", ss);
     }
